Add quoting/unquoting of HTML attributes containing JSX expressions
ParserHtmlCaseSensitive takes care of placing quotes around JSX
expression in attributes and also escaping some HTML entities (<,>,&).

Then, when the template is provided by the loadHtmltemplate command
we unquote it and pass it like this to TemplateDisplay.

With this we can use the same react component both in the live view
of grapsejs, generate the same component in the template with cases kept,
and then export the template and load it in a separate react component
as a JSX string turning again into a live react component.